fix(sync): float->short UB sweep — same class as the COB desync bug
Audit sweep of out-of-range float->short conversions in synced code
(IPathController stopTurnHeading / turn-rate clamps / return value,
HoverAirMoveType + GroundMoveType direct-control heading,
LuaSyncedMoveCtrl.SetHeading, LuaSyncedRead.GetFacingFromHeading).
All now truncate via int32: defined behavior, identical to what the
fleet's gcc-x86 builds compute.

+++ rts/Sim/MoveTypes/GroundMoveType.cpp
@@ -3316,8 +3316,8 @@ bool CGroundMoveType::UpdateDirectControl()
 		ChangeSpeed(0.0f, false, true);
 	}
 
-	if (unitCon.left ) { ChangeHeading(owner->heading + turnRate); turnSign =  1.0f; }
-	if (unitCon.right) { ChangeHeading(owner->heading - turnRate); turnSign = -1.0f; }
+	if (unitCon.left ) { ChangeHeading(short(int(owner->heading + turnRate))); turnSign =  1.0f; }
+	if (unitCon.right) { ChangeHeading(short(int(owner->heading - turnRate))); turnSign = -1.0f; }
 
 	// local client is controlling us
 	if (selfCon.GetControllee() == owner)
